% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_map.R
\name{cor_map}
\alias{cor_map}
\title{A function to implement a correlation heatmap including coefficients based on given numeric columns of a data frame.}
\usage{
cor_map(dataframe, num_col, col_scheme = "purpleorange")
}
\arguments{
\item{dataframe}{The data frame to be used for EDA.}

\item{num_col}{A vector of string of column names with numeric data from the data frame.}

\item{col_scheme}{The color scheme of the heatmap desired, can only be one of the following: 'purpleorange', 'blueorange', 'redblue'. Defaults to 'purpleorange'.}
}
\value{
ggplot object; A correlation heatmap plot with correlation coefficient labels based on the numeric columns specified by user.
}
\description{
A function to implement a correlation heatmap including coefficients based on given numeric columns of a data frame.
}
\examples{
data <- data.frame(
SepalLengthCm' = c(5.1, 4.9, 4.7), 
SepalWidthCm'= c(1.4, 1.4, 1.3), 
PetalWidthCm'= c(0.2, 0.1, 0.2), 
Species' = c('Iris-setosa','Iris-virginica', 'Iris-germanica')
)
numerical_columns <- c('SepalLengthCm','SepalWidthCm','PetalWidthCm')
 
cor_map(data, numerical_columns, col_scheme = 'purpleorange')

}
