% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputer.R
\name{imputer}
\alias{imputer}
\title{A function to implement imputation functionality for completing missing values.}
\usage{
imputer(df, strategy = "mean", fill_value = NULL)
}
\arguments{
\item{df}{list}

\item{strategy}{character}

\item{fill_value}{character}
}
\value{
list
}
\description{
A function to implement imputation functionality for completing missing values.
}
\examples{
df = tibble::tibble(
col1 = c(1, 2, NULL),
col2 = c(2, NULL, NULL)
)
imputer(df)
}
