% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.R
\name{outlier_identifier}
\alias{outlier_identifier}
\title{A function that identify by z-test with threshold of 2.5, and deal with outliers based on the method the user choose.}
\usage{
outlier_identifier(dataframe, columns = NULL, method = "trim")
}
\arguments{
\item{dataframe}{The target dataframe where the function is performed.}

\item{columns}{A vector of the target columns where the function needed to be performed. Defualt is NULL, the function will check all columns}

\item{method}{The method of dealing with outliers. Defualt is "trim"}
}
\value{
dataframe object; a dataframe which the outlier has already process by the chosen method
}
\description{
A function that identify by z-test with threshold of 2.5, and deal with outliers based on the method the user choose.
}
\examples{
data = data.frame(
SepalLengthCm = c(5.1, 4.9, 4.7, 5.4),
SepalWidthCm = c(1.4, 1.4, 20, 2.3),
PetalWidthCm = c(0.2, 0.2, 0.2, 0.9)
)
 
outlier_identifier(data)

}
