% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale}
\alias{scale}
\title{A function to scale features either by using standard scaler or minmax scaler method}
\usage{
scale(df, columns, scaler = "standard")
}
\arguments{
\item{df}{The data frame to be used for EDA.}

\item{columns}{A vector of string of column names with numeric data from the data frame.}

\item{scaler}{The method for scaling data and can only be one of the following: 'standard', 'minmax'}
}
\value{
The scaled dataframe for numerical features
}
\description{
A function to scale features either by using standard scaler or minmax scaler method
}
\examples{
data1 = data.frame(
"col1" = c(1, 2, 1),
"col2" = c(2, 2, 1)
)
scale(data1, c('col1', 'col2'))
}
